buscador-lucene/python/requirements.txt
pylucene==9.0.0
PyPDF2==3.0.1
python-docx==0.8.11
psycopg2-binary==2.9.5
python-dotenv==1.0.0
